+++ Packages/com.orchidseal.billboard/Editor/BaseBillboardEditor.cs
@@ -8,29 +8,42 @@ namespace OrchidSeal.Billboard.Editor
         protected bool showStencilOptions;
         private bool showTransformationOptions = true;
         
+        private static readonly int s_ButtonLinkHash = nameof(s_ButtonLinkHash).GetHashCode();
+        
         private static class Styles
         {
-            public static readonly GUIStyle linkIcon = new()
+            public static readonly GUIStyle linkIcon = new(EditorStyles.iconButton)
             {
                 fixedHeight = 20,
                 fixedWidth = 20,
             };
 
+            public static readonly Color linkIconFocusColor = new(0.23f, 0.47f, 0.73f);
+            public static readonly Color linkIconHoverColor = new(0.8f, 0.8f, 0.8f);
             public static readonly Color aboutBarColor = new(0.29f, 0.29f, 0.29f);
             public static int aboutBarHeight = 26;
             public static GUIStyle aboutHorizontalLayout = new(EditorStyles.inspectorFullWidthMargins)
             {
                 margin = new RectOffset(0, 0, 3, 1),
             };
-            
+            public static GUIStyle editorHeading = new()
+            {
+                alignment = TextAnchor.MiddleCenter,
+                fontSize = 14,
+                normal = new GUIStyleState
+                {
+                    textColor = new Color(0.75f, 0.75f, 0.75f),
+                },
+            };
             public static GUIStyle sectionVerticalLayout = new()
             {
                 margin = new RectOffset(0, 0, 0, 12),
             };
             
             // About
-            public static readonly GUIContent supportIcon = new (Resources.Load<Texture2D>("kofi_symbol"));
-            public static readonly GUIContent websiteIcon = new (Resources.Load<Texture2D>("Orchid Seal Logo"));
+            public static readonly GUIContent editorTitle = new ("OSP Billboard");
+            public static readonly GUIContent supportIcon = new (Resources.Load<Texture2D>("kofi_symbol"), "Ko-fi");
+            public static readonly GUIContent websiteIcon = new (Resources.Load<Texture2D>("Orchid Seal Logo"), "Orchid Seal");
             
             // Transformation
             public const string transformationFoldoutLabel = "Transformation";
@@ -39,7 +52,7 @@ namespace OrchidSeal.Billboard.Editor
             public static readonly GUIContent scaleLabel = new ("Scale");
             public static readonly GUIContent billboardModeLabel = new ("Billboard Mode");
             public static readonly GUIContent useNonUniformScaleLabel = new ("Use Non Uniform Object Scale");
-            public static readonly GUIContent keepConstantScalingLabel = new ("Keep Constant Scaling");
+            public static readonly GUIContent keepConstantScalingLabel = new ("Constant with Screen Size");
             public static readonly GUIContent constantScaleLabel = new ("Constant Scale");
             
             // Stencil
@@ -56,7 +69,19 @@ namespace OrchidSeal.Billboard.Editor
         private static bool ButtonLink(Rect rect, GUIContent content, GUIStyle style)
         {
             EditorGUIUtility.AddCursorRect(rect, MouseCursor.Link);
-            return GUI.Button(rect, content, style);
+            var id = GUIUtility.GetControlID(s_ButtonLinkHash, FocusType.Keyboard, rect);
+            var current = Event.current;
+            if (current.type == EventType.Repaint)
+            {
+                if (GUIUtility.keyboardControl == id) GUI.contentColor = Styles.linkIconFocusColor;
+                if (rect.Contains(current.mousePosition)) GUI.contentColor = Styles.linkIconHoverColor;
+            }
+            var result = ShaderGuiUtility.DrawControl(rect, id, content, style, false);
+            if (current.type == EventType.Repaint)
+            {
+                GUI.contentColor = Color.white;
+            }
+            return result;
         }
 
         protected static void AboutLinks()
@@ -82,6 +107,12 @@ namespace OrchidSeal.Billboard.Editor
             {
                 Application.OpenURL("https://orchidseal.com");
             }
+            buttonRect.x += 28;
+
+            var labelRect = new Rect(buttonRect);
+            labelRect.width = 120;
+            labelRect.x = Mathf.Max(Screen.width / 2 - labelRect.width / 2, buttonRect.x);
+            GUI.Label(labelRect, Styles.editorTitle, Styles.editorHeading);
         }
         
         protected void TransformationOptions(MaterialEditor materialEditor, MaterialProperty[] properties)