+++ Packages/com.orchidseal.billboard/Editor/LitBillboardEditor.cs
@@ -23,8 +23,8 @@ namespace OrchidSeal.Billboard.Editor
                 margin = new RectOffset(0, 0, 0, 12),
             };
             
-            // Blending option labels
-            public const string blendingFoldoutLabel = "Blending";
+            // Render option labels
+            public const string renderFoldoutLabel = "Rendering";
             public static readonly GUIContent renderModeLabel = new("Render Mode", "Opaque:\nCannot be seen through.\n\nCutout:\nCut holes in geometry by discarding any pixel whose combined alpha is below a cutoff threshold.\n\nTransparent:\nSmoothly blended transparency that uses the combined alpha values of pixels.\n\nPremultiply:\nTransparency using colors that are premultiplied with the alpha value, which can improve the appearance of soft edges.\n\nAdditive:\nGlowing transparency. Add the numbers for the two colors together.\n\nCustom:\nControl all blending settings separately.");
             public static readonly GUIContent useAlphaTestLabel = new GUIContent("Use Alpha Test");
             public static readonly GUIContent alphaCutoffLabel = new GUIContent("Alpha Cutoff");
@@ -52,7 +52,7 @@ namespace OrchidSeal.Billboard.Editor
         private static readonly int EmissionMap = Shader.PropertyToID("_EmissionMap");
 
         private bool showBaseOptions = true;
-        private bool showBlendingOptions = true;
+        private bool showRenderOptions = true;
         private bool showEmissionOptions;
 
         public override void OnGUI(MaterialEditor materialEditor, MaterialProperty[] properties)
@@ -60,7 +60,7 @@ namespace OrchidSeal.Billboard.Editor
             var targetMaterial = materialEditor.target as Material;
             
             AboutLinks();
-            BlendingOptions(materialEditor, properties, targetMaterial);
+            RenderOptions(materialEditor, properties, targetMaterial);
             BaseOptions(materialEditor, properties, targetMaterial);
             EmissionOptions(materialEditor, properties, targetMaterial);
             TransformationOptions(materialEditor, properties);
@@ -69,9 +69,9 @@ namespace OrchidSeal.Billboard.Editor
             materialEditor.EnableInstancingField();
         }
 
-        private void BlendingOptions(MaterialEditor materialEditor, MaterialProperty[] properties, Material targetMaterial)
+        private void RenderOptions(MaterialEditor materialEditor, MaterialProperty[] properties, Material targetMaterial)
         {
-            if (ShaderGuiUtility.FoldoutHeader(Styles.blendingFoldoutLabel, ref showBlendingOptions))
+            if (ShaderGuiUtility.FoldoutHeader(Styles.renderFoldoutLabel, ref showRenderOptions))
             {
                 EditorGUILayout.BeginVertical(Styles.sectionVerticalLayout);